Resolved code review comments from PR 256
