+++ src/tape/__init__.py
@@ -1,3 +1,4 @@
 from .analysis import *  # noqa
 from .ensemble import *  # noqa
 from .timeseries import *  # noqa
+from .ensemble_readers import *  # noqa
